+++ Nickel/Utilities/ManifestExt.cs
@@ -24,12 +24,15 @@ public static class ManifestExt
 			return new Error<string>($"`{nameof(IAssemblyModManifest.EntryPointType)}` value is invalid.");
 		if (!TryParseAssemblyReferences(out var assemblyReferences))
 			return new Error<string>($"`{nameof(IAssemblyModManifest.AssemblyReferences)}` value is invalid.");
+		if (!TryParseMethodsToStopInlining(out var methodsToStopInlining))
+			return new Error<string>($"`{nameof(IAssemblyModManifest.MethodsToStopInlining)}` value is invalid.");
 
 		return new AssemblyModManifest(manifest)
 		{
 			EntryPointAssembly = entryPointAssembly,
 			EntryPointType = entryPointType,
-			AssemblyReferences = assemblyReferences ?? []
+			AssemblyReferences = assemblyReferences ?? [],
+			MethodsToStopInlining = methodsToStopInlining ?? [],
 		};
 
 		bool TryParseEntryPointAssemblyFileName([MaybeNullWhen(false)] out string result)
@@ -64,6 +67,22 @@ public static class ManifestExt
 			if (!manifest.ExtensionData.TryGetValue(nameof(IAssemblyModManifest.AssemblyReferences), out var raw))
 				return true;
 			var nullableResult = JsonConvert.DeserializeObject<List<ModAssemblyReference>>(JsonConvert.SerializeObject(raw, settings), settings);
+			if (nullableResult is null)
+				return false;
+			result = nullableResult;
+			return true;
+		}
+
+		bool TryParseMethodsToStopInlining(out IReadOnlyList<StopInliningDefinition>? result)
+		{
+			var settings = new JsonSerializerSettings
+			{
+				ConstructorHandling = ConstructorHandling.AllowNonPublicDefaultConstructor
+			};
+			result = null;
+			if (!manifest.ExtensionData.TryGetValue(nameof(IAssemblyModManifest.MethodsToStopInlining), out var raw))
+				return true;
+			var nullableResult = JsonConvert.DeserializeObject<List<StopInliningDefinition>>(JsonConvert.SerializeObject(raw, settings), settings);
 			if (nullableResult is null)
 				return false;
 			result = nullableResult;